+++ lib/dx/rule.ex
@@ -1,14 +1,15 @@
 defmodule Dx.Rule do
-  @moduledoc """
-  Represents a rule, based on an individual record.
+  # Represents a rule, based on an individual record.
+  #
+  # ## `when`
+  # Nested map with data requirements and matching values.
+  #
+  # Lists mean one of the values must match.
+  #
+  # Negations can be expressed using `{:not, value}` or `{:not, [values]}`.
 
-  ## `when`
-  Nested map with data requirements and matching values.
+  @moduledoc false
 
-  Lists mean one of the values must match.
-
-  Negations can be expressed using `{:not, value}` or `{:not, [values]}`.
-  """
   use TypedStruct
 
   typedstruct do