+++ test/support/factories.ex
@@ -1,7 +1,7 @@
 defmodule Test.Support.Factories do
-  use Refinery, repo: Dx.Test.Repo
+  use Refactory, repo: Dx.Test.Repo
 
-  def refinement(Dx.Test.Schema.User, :default) do
+  def trait(Dx.Test.Schema.User, :default) do
     %{
       email: "alice@acme.org",
       first_name: "Alice",
@@ -9,19 +9,19 @@ defmodule Test.Support.Factories do
     }
   end
 
-  def refinement(Dx.Test.Schema.List, :default) do
+  def trait(Dx.Test.Schema.List, :default) do
     %{
       title: "My List"
     }
   end
 
-  def refinement(Dx.Test.Schema.ListTemplate, :default) do
+  def trait(Dx.Test.Schema.ListTemplate, :default) do
     %{
       title: "My List Template"
     }
   end
 
-  def refinement(Dx.Test.Schema.Task, :default) do
+  def trait(Dx.Test.Schema.Task, :default) do
     %{
       title: "My Task"
     }