+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2024-05-13
+
+- Configure dataloader (#21) - @ftes
+- Handle `:timeout` atom as error (#20) - @ftes
+
 ## [0.3.0] - 2022-07-10
 
 Initial release